+++ App/include/RayTracer/Abstraction/IShape.hpp
@@ -22,13 +22,13 @@ namespace RayTracer {
         public:
 
             virtual void setType(const ShapeType& type) = 0;
-            virtual void setRadius(float radius) = 0;
+            virtual void setRadius(int16_t radius) = 0;
             virtual void setMaterial(std::unique_ptr<AMaterial> material) = 0;
 
             [[nodiscard]] virtual ShapeType getType() const = 0;
             [[nodiscard]] virtual AMaterial* getMaterial() const = 0;
             [[nodiscard]] virtual Vector& getPosition() = 0;
-            [[nodiscard]] virtual float getRadius() const = 0;
+            [[nodiscard]] virtual int16_t getRadius() const = 0;
 
     }; // IShape
